+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased - NEOM U
-
+### Added
+- Add Messages section to Favorites [#37](https://github.com/rokmetro/neom-app/issues/37)
 ## [0.5.2] = 2025-01-17
 ### Fixed
 - Update User Directory preview styling [#32](https://github.com/rokmetro/neom-app/issues/32)